# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MechanicTranslator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  code?: string;
}

export interface WebhookAck {
  received: boolean;
}

export interface PriceRange {
  min: number;
  max: number;
  currency: string;
}

export type Urgency = (typeof Urgency)[keyof typeof Urgency];

export const Urgency = {
  Critical: "Critical",
  Soon: "Soon",
  Optional: "Optional",
} as const;

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  fair: "fair",
  high: "high",
  suspicious: "suspicious",
} as const;

export interface RepairItem {
  title: string;
  explanation: string;
  urgency: Urgency;
  fairPrice: PriceRange;
  quotedPrice?: number | null;
  flags: string[];
}

export interface TranslationResult {
  summary: string;
  verdict: Verdict;
  quotedTotal?: number | null;
  fairTotal: PriceRange;
  items: RepairItem[];
  redFlags: string[];
  negotiationScript: string;
}

export interface Translation {
  id: number;
  inputText: string;
  result: TranslationResult;
  createdAt: string;
}

export interface TranslateRequest {
  /**
   * @minLength 5
   * @maxLength 8000
   */
  inputText: string;
}

export type SubscriptionTier =
  (typeof SubscriptionTier)[keyof typeof SubscriptionTier];

export const SubscriptionTier = {
  free: "free",
  pro: "pro",
} as const;

export interface MeResponse {
  id: number;
  clerkUserId: string;
  email: string;
  subscriptionTier: SubscriptionTier;
  translationsUsed: number;
  freeLimit: number;
  translationsRemaining?: number | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface CheckoutSession {
  url?: string | null;
  configured: boolean;
  message: string;
}

export interface AdminStats {
  totalUsers: number;
  totalTranslations: number;
  freeUsers: number;
  proUsers: number;
  translationsLast7Days: number;
  signupsLast7Days: number;
}

export interface AdminUser {
  id: number;
  email: string;
  subscriptionTier: SubscriptionTier;
  translationsUsed: number;
  createdAt: string;
}