# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MechanicTranslator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Start a Pro subscription checkout
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Billing not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /billing/webhook:
    post:
      operationId: stripeWebhook
      tags: [billing]
      summary: Stripe webhook receiver
      description: Receives Stripe events. Raw body required for Stripe signature verification — do not parse as JSON.
      responses:
        "200":
          description: Event received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
        "400":
          description: Invalid or missing Stripe signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    WebhookAck:
      type: object
      properties:
        received:
          type: boolean
      required: [received]
...
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
          nullable: true
        configured:
          type: boolean
        message:
          type: string
      required: [configured, message]
...